+++ activerecord/test/cases/tasks/postgresql_rake_test.rb
@@ -195,21 +195,54 @@ module ActiveRecord
         'adapter'  => 'postgresql',
         'database' => 'my-app-db'
       }
+      @filename = "awesome-file.sql"
 
       ActiveRecord::Base.stubs(:connection).returns(@connection)
       ActiveRecord::Base.stubs(:establish_connection).returns(true)
       Kernel.stubs(:system)
+      File.stubs(:open)
     end
 
     def test_structure_dump
-      filename = "awesome-file.sql"
-      Kernel.expects(:system).with("pg_dump -i -s -x -O -f #{filename}  my-app-db").returns(true)
-      @connection.expects(:schema_search_path).returns("foo")
+      Kernel.expects(:system).with("pg_dump -i -s -x -O -f #{@filename}  my-app-db").returns(true)
 
-      ActiveRecord::Tasks::DatabaseTasks.structure_dump(@configuration, filename)
-      assert File.exist?(filename)
+      ActiveRecord::Tasks::DatabaseTasks.structure_dump(@configuration, @filename)
+    end
+
+    def test_structure_dump_with_schema_search_path
+      @configuration['schema_search_path'] = 'foo,bar'
+
+      Kernel.expects(:system).with("pg_dump -i -s -x -O -f #{@filename} --schema=foo --schema=bar my-app-db").returns(true)
+
+      ActiveRecord::Tasks::DatabaseTasks.structure_dump(@configuration, @filename)
+    end
+
+    def test_structure_dump_with_schema_search_path_and_dump_schemas_all
+      @configuration['schema_search_path'] = 'foo,bar'
+
+      Kernel.expects(:system).with("pg_dump -i -s -x -O -f #{@filename}  my-app-db").returns(true)
+
+      with_dump_schemas(:all) do
+        ActiveRecord::Tasks::DatabaseTasks.structure_dump(@configuration, @filename)
+      end
+    end
+
+    def test_structure_dump_with_dump_schemas_string
+      Kernel.expects(:system).with("pg_dump -i -s -x -O -f #{@filename} --schema=foo --schema=bar my-app-db").returns(true)
+
+      with_dump_schemas('foo,bar') do
+        ActiveRecord::Tasks::DatabaseTasks.structure_dump(@configuration, @filename)
+      end
+    end
+
+    private
+
+    def with_dump_schemas(value, &block)
+      old_dump_schemas = ActiveRecord::Base.dump_schemas
+      ActiveRecord::Base.dump_schemas = value
+      yield
     ensure
-      FileUtils.rm(filename)
+      ActiveRecord::Base.dump_schemas = old_dump_schemas
     end
   end