+++ activerecord/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Add `config.active_record.dump_schemas` to fix `db:structure:dump`
+    when using schema_search_path and PostgreSQL extensions.
+
+    Fixes #17157.
+
+    *Ryan Wallace*
+
 *   Renaming `use_transactional_fixtures` to `use_transactional_tests` for clarity.
 
     Fixes #18864.